server: add option to disable primary gateway fallback when using mesh gateway wan federation
If after initial federation via mesh gateways the network connectivity
between the mesh gateways in the primary datacenter and the servers in a
secondary datacenter is voluntarily severed, setting this option will
prevent attempts to use the primary directly.

Failure to do this could lead to broken federation if the local gateawys
ever cease to function.